backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.2.0
torchvision==0.17.0
Pillow==8.3.1 